+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Yii Framework 2 apidoc extension Change Log
 - Bug #12: Do not publish PHP files for `jssearch.js` asset (cebe)
 - Bug #42: Fixed stopword filter in JS search index, which resulted in empty results for some words like `sort` (cebe)
 - Bug #62: Make `@inheritdoc` tag more robust (cebe, sasha-ch)
+- Bug #67: Use multibyte compatible function for `ucfirst()` in descriptions (cebe, samdark)
 - Bug #68: Fixed crash on empty type in PHPdoc (cebe, itnelo)
 - Bug #76: Fixed broken links with external urls (CedricYii)
 - Bug #79: Fixed crash due to missing encoding specified in `mb_*` functions (cebe, dingzhihao)